feat: ft-wave1-cursor-golden-success-003 - Consume migration-ledger rows and close verification gates
