+++ docs/internal/baselines/functional-undocumented-tests.json
@@ -536,10 +536,6 @@
     },
     {
       "file": "providers/cli_template_resolution_long_test.go",
-      "name": "TestTemplateTests_ScriptWrapCursorResolvesWorkstationExecutionTemplates"
-    },
-    {
-      "file": "providers/cli_template_resolution_long_test.go",
       "name": "TestTemplateTests_ScriptWrapDropsResourceTokensFromWorkstationTemplates"
     },
     {
@@ -573,18 +569,6 @@
     {
       "file": "providers/codex_worktree_workstation_test.go",
       "name": "TestCodexWorktreeWorkstationDispatch_MaterializesCheckoutAndOmitsCLIWorktreeFlag"
-    },
-    {
-      "file": "providers/cursor_provider_command_test.go",
-      "name": "TestCursorProviderCommand_DispatchesAgentWithRenderedPrompt"
-    },
-    {
-      "file": "providers/cursor_provider_command_test.go",
-      "name": "TestCursorProviderCommand_PublicModelProviderEnumRoutesToAgent"
-    },
-    {
-      "file": "providers/cursor_provider_command_test.go",
-      "name": "TestCursorProviderCommand_SkipPermissionsPassesForceFlag"
     },
     {
       "file": "providers/mock_workers_agent_test.go",
