+++ docs/internal/baselines/package-structure-baseline.json
@@ -1034,11 +1034,6 @@
     },
     {
       "rule": "functional-test-missing-subsection",
-      "filePath": "tests/functional/providers/cursor_provider_command_test.go",
-      "target": "providers"
-    },
-    {
-      "rule": "functional-test-missing-subsection",
       "filePath": "tests/functional/providers/doc.go",
       "target": "providers"
     },
